+++ examples/resources/nxos_pim_anycast_rp_peer/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_pim_anycast_rp_peer.example "sys/pim/inst/dom-[default]/acastrpfunc/peer-[10.1.1.1/32]-peer-[20.1.1.1/32]"